apiKey: migrate to separate types file
